+++ NEWS.md
@@ -42,6 +42,7 @@
   is deprecated and forwards to `only_usd` with a warning.
 - Added `test-coverage.yaml` workflow and badge to README.
 - Added tests to get coverage to 100% (excl. `set_wrds_credentials()`).
+- Removed `arrow`, `glue`, and `stringr` dependencies and added `nanoparquet`.
 
 # tidyfinance 0.5.0